chore(funnels): move funnel hog-charts viz into product_analytics
Continues the migration started in #59674. Moves the three flag-gated hog-charts
funnel viz components plus their shared meta out of `frontend/src/scenes/funnels/viz/`
and into `products/product_analytics/frontend/insights/funnels/`.

Same shape as the trends slice: intra-product imports use relative paths;
back-pointers into still-legacy funnels modules (funnelDataLogic,
funnelPersonsModalLogic, funnelUtils, FunnelTooltip, FunnelBarVertical/StepLegend)
use absolute `scenes/funnels/*` aliases that dissolve as the rest of the scene
follows. `Funnel.tsx` lazy-imports the moved charts from their new paths and
wraps the viz in a Suspense boundary with WrappingLoadingSkeleton, matching
Trends.tsx.

Generated-By: PostHog Code
Task-Id: 329f4a6d-da4c-47e9-bdf4-5af9131089a5
