+++ products/product_analytics/frontend/insights/funnels/FunnelHistogramChart/funnelHistogramTransforms.ts
