+++ products/product_analytics/frontend/insights/funnels/FunnelLineChart/FunnelLineChart.tsx
@@ -5,6 +5,8 @@ import { useMemo, type ErrorInfo } from 'react'
 import { buildTheme } from 'lib/charts/utils/theme'
 import { TimeSeriesLineChart } from 'lib/hog-charts'
 import type { PointClickData, TimeSeriesLineChartConfig, TooltipConfig, TooltipContext } from 'lib/hog-charts'
+import { funnelDataLogic } from 'scenes/funnels/funnelDataLogic'
+import { funnelPersonsModalLogic } from 'scenes/funnels/funnelPersonsModalLogic'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import type { SeriesDatum } from 'scenes/insights/InsightTooltip/insightTooltipUtils'
 import { teamLogic } from 'scenes/teamLogic'
@@ -18,10 +20,7 @@ import { propertyDefinitionsModel } from '~/models/propertyDefinitionsModel'
 import { isFunnelsQuery } from '~/queries/utils'
 import { ChartParams, type FunnelStepWithConversionMetrics } from '~/types'
 
-import { AnnotationsLayer } from 'products/product_analytics/frontend/insights/trends/shared/AnnotationsLayer'
-
-import { funnelDataLogic } from '../../funnelDataLogic'
-import { funnelPersonsModalLogic } from '../../funnelPersonsModalLogic'
+import { AnnotationsLayer } from '../../trends/shared/AnnotationsLayer'
 import type { FunnelSeriesMeta } from '../shared/funnelSeriesMeta'
 import { buildFunnelLineSeries, buildFunnelLineTimeSeriesConfig, type IndexedFunnelStep } from './funnelChartTransforms'
 import { FunnelLineTooltip } from './FunnelLineTooltip'
