+++ products/product_analytics/frontend/insights/funnels/FunnelLineChart/funnelChartTransforms.test.ts
