+++ products/product_analytics/frontend/insights/funnels/shared/funnelSeriesMeta.ts
